+++ example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # example
 
+## 0.5.1
+
+### Patch Changes
+
+- Improve cross selection
+- Updated dependencies [undefined]
+  - @hexx/editor@0.6.1
+
 ## 0.5.0
 
 ### Minor Changes